refactyor: ユーザーの作品情報表示のファイル分割に際する変更
